+++ app/controlplane/pkg/authz/middleware/middleware.go
@@ -22,8 +22,8 @@ import (
 
 	errorsAPI "github.com/go-kratos/kratos/v2/errors"
 
-	"github.com/chainloop-dev/chainloop/app/controlplane/internal/authz"
 	"github.com/chainloop-dev/chainloop/app/controlplane/internal/usercontext"
+	"github.com/chainloop-dev/chainloop/app/controlplane/pkg/authz"
 	"github.com/go-kratos/kratos/v2/log"
 	"github.com/go-kratos/kratos/v2/middleware"
 	"github.com/go-kratos/kratos/v2/transport"
